refactor(config): 将配置管理逻辑迁移到独立模块
重构配置管理代码，将原本分散在__init__.py中的配置加载、验证和边界控制逻辑提取到新的loader.py模块
移除不再需要的全局配置变量和设置函数，统一通过loader模块管理配置
